ingestion/requirements.txt
pykalshi>=0.1.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
alembic>=1.13
